+++ packages/grafana-ui/src/components/uPlot/types.ts
@@ -1,5 +1,5 @@
 import React from 'react';
-import uPlot, { Options, Hooks } from 'uplot';
+import uPlot, { Options, Hooks, AlignedData } from 'uplot';
 import { DataFrame, TimeRange, TimeZone } from '@grafana/data';
 import { UPlotConfigBuilder } from './config/UPlotConfigBuilder';
 
@@ -17,7 +17,9 @@ export interface PlotPluginProps {
 }
 
 export interface PlotProps {
-  data: DataFrame;
+  data: AlignedData;
+  // TODO: left temporarily for Tooltip to work, to be removed
+  dataFrame: DataFrame;
   timeRange: TimeRange;
   timeZone: TimeZone;
   width: number;
